refactor: add DataSourceInfo::ParquetSource.
